+++ package-dev/Plugins/iOS/SentryNativeBridge.m
@@ -138,7 +138,7 @@ static inline NSString *_NSStringOrNil(const char *value)
     return value ? [NSString stringWithUTF8String:value] : nil;
 }
 
-static inline NSNumber *_NSNumberOrNil(int32_t value) 
+static inline NSNumber *_NSNumberOrNil(int32_t value)
 {
     return value == 0 ? nil : @(value);
 }
@@ -183,6 +183,7 @@ void SentryNativeBridgeWriteScope( // clang-format off
     const char *GpuVendorId,
     int8_t GpuMultiThreadedRendering,
     const char *GpuGraphicsShaderLevel,
+    const char *EditorVersion,
     const char *UnityInstallMode,
     const char *UnityTargetFrameRate,
     const char *UnityCopyTextureSupport,
@@ -210,6 +211,7 @@ void SentryNativeBridgeWriteScope( // clang-format off
         [scope performSelector:@selector(setContextValue:forKey:) withObject:gpu withObject:@"gpu"];
 
         NSMutableDictionary *unity = [[NSMutableDictionary alloc] init];
+        unity[@"editor_version"] = _NSStringOrNil(EditorVersion);
         unity[@"install_mode"] = _NSStringOrNil(UnityInstallMode);
         unity[@"target_frame_rate"] = _NSStringOrNil(UnityTargetFrameRate);
         unity[@"copy_texture_support"] = _NSStringOrNil(UnityCopyTextureSupport);