+++ package-dev/Plugins/macOS/SentryNativeBridge.m
@@ -289,6 +289,7 @@ void SentryNativeBridgeWriteScope( // clang-format off
     const char *GpuVendorId,
     int8_t GpuMultiThreadedRendering,
     const char *GpuGraphicsShaderLevel,
+    const char *EditorVersion,
     const char *UnityInstallMode,
     const char *UnityTargetFrameRate,
     const char *UnityCopyTextureSupport,
@@ -316,6 +317,7 @@ void SentryNativeBridgeWriteScope( // clang-format off
         [scope performSelector:@selector(setContextValue:forKey:) withObject:gpu withObject:@"gpu"];
 
         NSMutableDictionary *unity = [[NSMutableDictionary alloc] init];
+        unity[@"editor_version"] = _NSStringOrNil(EditorVersion);
         unity[@"install_mode"] = _NSStringOrNil(UnityInstallMode);
         unity[@"target_frame_rate"] = _NSStringOrNil(UnityTargetFrameRate);
         unity[@"copy_texture_support"] = _NSStringOrNil(UnityCopyTextureSupport);