+++ src/Sentry.Unity.Native/NativeContextWriter.cs
@@ -33,6 +33,7 @@ namespace Sentry.Unity.Native
             string? GpuVendorId,
             bool? GpuMultiThreadedRendering,
             string? GpuGraphicsShaderLevel,
+            string? EditorVersion,
             string? UnityInstallMode,
             string? UnityTargetFrameRate,
             string? UnityCopyTextureSupport,
@@ -74,11 +75,11 @@ namespace Sentry.Unity.Native
                 GpuGraphicsShaderLevel);
 
             CWUtil.WriteUnity(
+                EditorVersion,
                 UnityInstallMode,
                 UnityTargetFrameRate,
                 UnityCopyTextureSupport,
                 UnityRenderingThreadingMode);
         }
-
     }
 }